shapes-backend/pyproject.toml
[project]
name = "shapes-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "gradio>=5.31.0",
    "lightning>=2.5.1.post0",
    "mlflow>=2.22.0",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.3",
]
